Create cart context
Because the ProductPage component is dynamically
updated based on routing, the array that was
holding the cart data was being reset every time
a different product was being added to the cart.

I had to create a context for the cart, along with
a custom hook. This way, the cart state can be
managed across the entire application.