Find cart quantities
Added a function in Cart component that takes
the cartContents array and calculates the
quantities of each item.
Then used this new object of cart items and their
quantites to render the items in the cart.
Will be adding buttons to be able to increase
quantities and delete items from the cart.