Update addToCart function
Combined the addToCart function and the
getCartWithQuantities function. The addToCart
function now adds the item to cart while also
adding the quantity key for each item.
If the item already exists in the cart and the
same product is added again, the quantity is
increased by 1.

Also added two buttons to the the quantity for
each item. These buttons will allow the user to
increase or decrease the quantity